barra de ação, tempo de tiro.
git-svn-id: https://timeline-chronicles.googlecode.com/svn/trunk@40 441dad9d-d751-dd0b-1c3e-f44f5c70b050

+++ Timeline Chronicles/src/Timeline/Entidade/Atributo/Posicao.java
@@ -9,8 +9,8 @@ package Timeline.Entidade.Atributo;
  * @author Desenv01
  */
 public class Posicao{
-    private int x;
-    private int y;
+    private double x;
+    private double y;
 
     public Posicao(){
         
@@ -24,28 +24,28 @@ public class Posicao{
     /**
      * @return the x
      */
-    public int getX() {
+    public double  getX() {
         return x;
     }
 
     /**
      * @param x the x to set
      */
-    public void setX(int x) {
+    public void setX(double x) {
         this.x = x;
     }
 
     /**
      * @return the y
      */
-    public int getY() {
+    public double getY() {
         return y;
     }
 
     /**
      * @param y the y to set
      */
-    public void setY(int y) {
+    public void setY(double y) {
         this.y = y;
     }
